Move processJob definition to plugin
