+++ imports/plugins/included/jobcontrol/server/index.js
@@ -1,5 +1,7 @@
 import "./jobs/exchangerates";
 import "./jobs/cleanup";
 import cleanupJob from "./jobs/cleanup";
+import fetchRateJobs from "./jobs/exchangerates";
 
 cleanupJob();
+fetchRateJobs();
