+++ server/startup/index.js
@@ -1,6 +1,5 @@
 import Accounts from "./accounts";
 import i18n from "./i18n";
-import Jobs from "./jobs";
 import Load from "./load-data";
 import Packages from "./packages";
 import Registry from "./registry";
@@ -9,7 +8,6 @@ import Init from "./init";
 export default function () {
   Accounts();
   i18n();
-  Jobs();
   Load();
   Packages();
   Registry();
